+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improved `TaskResultExt`. Now takes a `FnOnce() -> Context` instead of `&str`.
 
+### Fixed
+
+- Ensure a `Signature` inherits default `TaskOptions` from the corresponding `Task` and `Celery` app.
+
 ## v0.4.0-rc3 - 2020-09-09
 
 ### Added